[firebase-interfacing] Renamed FireBase to Firebase
