--- src/components/FireBase/Storage/api.js
@@ -1,5 +1,0 @@
-import { firebase } from '@react-native-firebase/storage';
-
-const FireBaseStorage = firebase.storage();
-
-export default FireBaseStorage;
