+++ src/components/Firebase/Storage/api.js
@@ -1,0 +1,5 @@
+import { firebase } from '@react-native-firebase/storage';
+
+const FirebaseStorage = firebase.storage();
+
+export default FirebaseStorage;
